funcionalidad de cliente

+++ src/app/components/navbar/navbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { Router, RouterLink, RouterLinkActive } from '@angular/router';
 import { IUser } from '../../interfaces/iuser';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-navbar',
@@ -13,12 +14,30 @@ export class NavbarComponent {
   router = inject(Router);
   userRole: string;
   user: IUser;
+
   constructor() {
-    this.user = JSON.parse(localStorage.getItem('user')!); // Asumiendo que el rol del usuario está almacenado en localStorage
+    this.user = JSON.parse(localStorage.getItem('user')!);
     this.userRole = this.user.rol || '';
   }
+
   logout() {
-    localStorage.clear();
-    this.router.navigate(['/login']);
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Vas a cerrar sesión.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, cerrar sesión',
+      cancelButtonText: 'Cancelar',
+      customClass: {
+        confirmButton: 'btn btn-secondary',
+        cancelButton: 'btn btn-outline-secondary ms-2',
+      },
+      buttonsStyling: false,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        this.router.navigate(['/login']);
+      }
+    });
   }
 }
